TornadoDemo/requirements.txt
xlwt
pycrypto
lxml
beautifulsoup4
qrcode
xmltodict
dicttoxml
bcrypt
bidict
intervaltree
intervals
redis
tornado
wtforms
peewee
pymysql
aiomysql
PyJWT
aiofiles
pillow
peewee-async